Simplify `iced_futures::Command` implementation
